+++ rx-preferences/src/main/java/com/f2prateek/rx/preferences2/RxSharedPreferences.java
@@ -1,20 +1,20 @@
-package com.f2prateek.rx.preferences;
+package com.f2prateek.rx.preferences2;
 
-import android.annotation.TargetApi;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.annotation.RequiresApi;
+import io.reactivex.Observable;
+import io.reactivex.ObservableEmitter;
+import io.reactivex.ObservableOnSubscribe;
+import io.reactivex.functions.Cancellable;
 import java.util.Collections;
 import java.util.Set;
-import rx.Observable;
-import rx.Subscriber;
-import rx.functions.Action0;
-import rx.subscriptions.Subscriptions;
 
 import static android.os.Build.VERSION_CODES.HONEYCOMB;
-import static com.f2prateek.rx.preferences.Preconditions.checkNotNull;
+import static com.f2prateek.rx.preferences2.Preconditions.checkNotNull;
 
 /** A factory for reactive {@link Preference} objects. */
 public final class RxSharedPreferences {
@@ -22,6 +22,7 @@ public final class RxSharedPreferences {
   private static final Integer DEFAULT_INTEGER = 0;
   private static final Boolean DEFAULT_BOOLEAN = false;
   private static final Long DEFAULT_LONG = 0L;
+  private static final String DEFAULT_STRING = "";
 
   /** Create an instance of {@link RxSharedPreferences} for {@code preferences}. */
   @CheckResult @NonNull
@@ -35,22 +36,22 @@ public final class RxSharedPreferences {
 
   private RxSharedPreferences(final SharedPreferences preferences) {
     this.preferences = preferences;
-    this.keyChanges = Observable.create(new Observable.OnSubscribe<String>() {
-      @Override public void call(final Subscriber<? super String> subscriber) {
+    this.keyChanges = Observable.create(new ObservableOnSubscribe<String>() {
+      @Override public void subscribe(final ObservableEmitter<String> emitter) throws Exception {
         final OnSharedPreferenceChangeListener listener = new OnSharedPreferenceChangeListener() {
           @Override
           public void onSharedPreferenceChanged(SharedPreferences preferences, String key) {
-            subscriber.onNext(key);
+            emitter.onNext(key);
           }
         };
 
-        preferences.registerOnSharedPreferenceChangeListener(listener);
-
-        subscriber.add(Subscriptions.create(new Action0() {
-          @Override public void call() {
+        emitter.setCancellable(new Cancellable() {
+          @Override public void cancel() throws Exception {
             preferences.unregisterOnSharedPreferenceChangeListener(listener);
           }
-        }));
+        });
+
+        preferences.registerOnSharedPreferenceChangeListener(listener);
       }
     }).share();
   }
@@ -63,26 +64,20 @@ public final class RxSharedPreferences {
 
   /** Create a boolean preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public Preference<Boolean> getBoolean(@NonNull String key, @Nullable Boolean defaultValue) {
+  public Preference<Boolean> getBoolean(@NonNull String key, @NonNull Boolean defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, BooleanAdapter.INSTANCE, keyChanges);
-  }
-
-  /** Create an enum preference for {@code key}. Default is {@code null}. */
-  @CheckResult @NonNull
-  public <T extends Enum<T>> Preference<T> getEnum(@NonNull String key,
-      @NonNull Class<T> enumClass) {
-    return getEnum(key, null, enumClass);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, BooleanAdapter.INSTANCE, keyChanges);
   }
 
   /** Create an enum preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public <T extends Enum<T>> Preference<T> getEnum(@NonNull String key, @Nullable T defaultValue,
+  public <T extends Enum<T>> Preference<T> getEnum(@NonNull String key, @NonNull T defaultValue,
       @NonNull Class<T> enumClass) {
     checkNotNull(key, "key == null");
+    checkNotNull(defaultValue, "defaultValue == null");
     checkNotNull(enumClass, "enumClass == null");
-    Preference.Adapter<T> adapter = new EnumAdapter<>(enumClass);
-    return new Preference<>(preferences, key, defaultValue, adapter, keyChanges);
+    return new RealPreference<>(preferences, key, defaultValue, new EnumAdapter<>(enumClass), keyChanges);
   }
 
   /** Create a float preference for {@code key}. Default is {@code 0}. */
@@ -93,9 +88,10 @@ public final class RxSharedPreferences {
 
   /** Create a float preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public Preference<Float> getFloat(@NonNull String key, @Nullable Float defaultValue) {
+  public Preference<Float> getFloat(@NonNull String key, @NonNull Float defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, FloatAdapter.INSTANCE, keyChanges);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, FloatAdapter.INSTANCE, keyChanges);
   }
 
   /** Create an integer preference for {@code key}. Default is {@code 0}. */
@@ -107,9 +103,10 @@ public final class RxSharedPreferences {
 
   /** Create an integer preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public Preference<Integer> getInteger(@NonNull String key, @Nullable Integer defaultValue) {
+  public Preference<Integer> getInteger(@NonNull String key, @NonNull Integer defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, IntegerAdapter.INSTANCE, keyChanges);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, IntegerAdapter.INSTANCE, keyChanges);
   }
 
   /** Create a long preference for {@code key}. Default is {@code 0}. */
@@ -121,54 +118,59 @@ public final class RxSharedPreferences {
 
   /** Create a long preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public Preference<Long> getLong(@NonNull String key, @Nullable Long defaultValue) {
+  public Preference<Long> getLong(@NonNull String key, @NonNull Long defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, LongAdapter.INSTANCE, keyChanges);
-  }
-
-  /** Create a preference of type {@code T} for {@code key}. Default is {@code null}. */
-  @CheckResult @NonNull
-  public <T> Preference<T> getObject(@NonNull String key, @NonNull Preference.Adapter<T> adapter) {
-    return getObject(key, null, adapter);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, LongAdapter.INSTANCE, keyChanges);
   }
 
   /**
    * Create a preference for type {@code T} for {@code key} with a default of {@code defaultValue}.
    */
-  @CheckResult @NonNull
-  public <T> Preference<T> getObject(@NonNull String key, @Nullable T defaultValue,
-      @NonNull Preference.Adapter<T> adapter) {
+  @CheckResult @NonNull public <T> Preference<T> getObject(@NonNull String key,
+      @NonNull T defaultValue, @NonNull Preference.Converter<T> converter) {
     checkNotNull(key, "key == null");
-    checkNotNull(adapter, "adapter == null");
-    return new Preference<>(preferences, key, defaultValue, adapter, keyChanges);
+    checkNotNull(defaultValue, "defaultValue == null");
+    checkNotNull(converter, "converter == null");
+    return new RealPreference<>(preferences, key, defaultValue,
+        new ConverterAdapter<>(converter), keyChanges);
   }
 
-  /** Create a string preference for {@code key}. Default is {@code null}. */
+  /** Create a string preference for {@code key}. Default is {@code ""}. */
   @CheckResult @NonNull
   public Preference<String> getString(@NonNull String key) {
-    return getString(key, null);
+    return getString(key, DEFAULT_STRING);
   }
 
   /** Create a string preference for {@code key} with a default of {@code defaultValue}. */
   @CheckResult @NonNull
-  public Preference<String> getString(@NonNull String key, @Nullable String defaultValue) {
+  public Preference<String> getString(@NonNull String key, @NonNull String defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, StringAdapter.INSTANCE, keyChanges);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, StringAdapter.INSTANCE, keyChanges);
   }
 
-  /** Create a string set preference for {@code key}. Default is an empty set. */
-  @TargetApi(HONEYCOMB)
+  /**
+   * Create a string set preference for {@code key}. Default is an empty set. Note that returned set
+   * value will always be unmodifiable.
+   */
+  @RequiresApi(HONEYCOMB)
   @CheckResult @NonNull
   public Preference<Set<String>> getStringSet(@NonNull String key) {
     return getStringSet(key, Collections.<String>emptySet());
   }
 
   /** Create a string set preference for {@code key} with a default of {@code defaultValue}. */
-  @TargetApi(HONEYCOMB)
+  @RequiresApi(HONEYCOMB)
   @CheckResult @NonNull
   public Preference<Set<String>> getStringSet(@NonNull String key,
       @NonNull Set<String> defaultValue) {
     checkNotNull(key, "key == null");
-    return new Preference<>(preferences, key, defaultValue, StringSetAdapter.INSTANCE, keyChanges);
+    checkNotNull(defaultValue, "defaultValue == null");
+    return new RealPreference<>(preferences, key, defaultValue, StringSetAdapter.INSTANCE, keyChanges);
+  }
+  
+  public void clear() {
+    preferences.edit().clear().apply();
   }
 }