+++ rx-preferences/src/main/java/com/f2prateek/rx/preferences2/StringAdapter.java
@@ -1,13 +1,15 @@
-package com.f2prateek.rx.preferences;
+package com.f2prateek.rx.preferences2;
 
 import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
 
-final class StringAdapter implements Preference.Adapter<String> {
+final class StringAdapter implements RealPreference.Adapter<String> {
   static final StringAdapter INSTANCE = new StringAdapter();
 
   @Override public String get(@NonNull String key, @NonNull SharedPreferences preferences) {
-    return preferences.getString(key, null);
+    String value = preferences.getString(key, null);
+    assert value != null; // Not called unless key is present.
+    return value;
   }
 
   @Override public void set(@NonNull String key, @NonNull String value,